+++ lib/argparser.rb
@@ -3,6 +3,7 @@ require 'stringio'
 
 require 'argparser/version'
 require 'argparser/tools'
+require 'argparser/argument'
 require 'argparser/option'
-require 'argparser/default_parser'
+require 'argparser/errors'
 require 'argparser/arg_parser'